backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
requests
beautifulsoup4
yfinance
